requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
python-dotenv==1.1.0
pymongo==4.13.2
jinja2==3.1.6
python-multipart==0.0.20
certifi==2025.6.15